requirements.txt
streamlit
pandas
beautifulsoup4
requests
docling.document_converter